LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-child-of-head.tentative_designMode=on&white-space=pre-expected.txt



FAIL insertParagraph in empty <title style="display:block;white-space:pre"> should not split it assert_equals: 2 preformatted linefeeds should be inserted expected "<head><title>\n\n</title></head><body><br></body>" but got "<head><title></title></head><body><br></body>"
FAIL insertParagraph in <title style="display:block;white-space:pre"> containing text should not split it assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>ab\ncd</title></head><body><br></body>" but got "<head><title>abcd</title></head><body><br></body>"
FAIL insertParagraph in empty <style style="display:block;white-space:pre"> should not split it assert_equals: 2 preformatted linefeeds should be inserted expected "<head><title>iframe</title><style>\n\n</style></head><body><br></body>" but got "<head><title>iframe</title><style></style></head><body><br></body>"
FAIL insertParagraph in <style style="display:block;white-space:pre"> containing text should not split it assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>iframe</title><style>ab\ncd</style></head><body><br></body>" but got "<head><title>iframe</title><style>abcd</style></head><body><br></body>"
FAIL insertParagraph in empty <script style="display:block;white-space:pre"> should not split it assert_equals: 2 preformatted linefeeds should be inserted expected "<head><title>iframe</title><script>\n\n</script></head><body><br></body>" but got "<head><title>iframe</title><script></script></head><body><br></body>"
FAIL insertParagraph in <script style="display:block;white-space:pre"> containing text should not split it assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>iframe</title><script>// ab\n// cd</script></head><body><br></body>" but got "<head><title>iframe</title><script>// ab// cd</script></head><body><br></body>"
FAIL insertParagraph in empty <div style="white-space:pre"> in the <head> should split the <div> assert_in_array: The <div> should be split value "<head><title>iframe</title><div style=\"white-space:pre\"></div></head><body><br></body>" not in array ["<head><title>iframe</title><div style=\"white-space:pre\">\n</div><div style=\"white-space:pre\">\n</div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:pre\"><br></div><div style=\"white-space:pre\"><br></div></head><body><br></body>"]
FAIL insertParagraph in <div style="white-space:pre"> (containing only a <br>) in the <head> should split the <div> element assert_in_array: The <div> should be split value "<head><title>iframe</title><div style=\"white-space:pre\"><br></div></head><body><br></body>" not in array ["<head><title>iframe</title><div style=\"white-space:pre\">\n</div><div style=\"white-space:pre\">\n</div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:pre\"><br></div><div style=\"white-space:pre\">\n</div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:pre\">\n</div><div style=\"white-space:pre\"><br></div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:pre\"><br></div><div style=\"white-space:pre\"><br></div></head><body><br></body>"]
FAIL insertParagraph in <div style="white-space:pre"> (containing text) in the <head> should split the <div> element assert_in_array: The <div> should be split value "<head><title>iframe</title><div style=\"white-space:pre\">abcd</div></head><body><br></body>" not in array ["<head><title>iframe</title><div style=\"white-space:pre\">ab</div><div style=\"white-space:pre\">cd</div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:pre\">ab</div><div style=\"white-space:pre\">cd<br></div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:pre\">ab<br></div><div style=\"white-space:pre\">cd</div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:pre\">ab<br></div><div style=\"white-space:pre\">cd<br></div></head><body><br></body>"]
